--- src/Exercism.Analyzers.CSharp/SolutionCommentParameter.cs
@@ -1,3 +1,0 @@
-namespace Exercism.Analyzers.CSharp;
-
-internal record SolutionCommentParameter(string Key, string Value);
